refactor: use sindresorhus/defaults to simplify the code base
